mgmt, local generation for Hyperscale Forward Migration progress reporting changes in DatabaseOperations API package-composite-v5 (#46304)
* generate codes

* fix compilation failure

* update changelog

* update version_client.txt

* update pom.xml

* update README.md

* fix test cases

* fix test case

* update assets.json

* Disabled some test cases for sql

* add `<revapi.skip>true</revapi.skip>` for fix ci analyze exception

+++ sdk/resourcemanager/azure-resourcemanager-sql/src/main/java/com/azure/resourcemanager/sql/fluent/JobStepExecutionsClient.java
@@ -43,8 +43,8 @@ public interface JobStepExecutionsClient {
     @ServiceMethod(returns = ReturnType.COLLECTION)
     PagedFlux<JobExecutionInner> listByJobExecutionAsync(String resourceGroupName, String serverName,
         String jobAgentName, String jobName, UUID jobExecutionId, OffsetDateTime createTimeMin,
-        OffsetDateTime createTimeMax, OffsetDateTime endTimeMin, OffsetDateTime endTimeMax, Boolean isActive, Long skip,
-        Long top);
+        OffsetDateTime createTimeMax, OffsetDateTime endTimeMin, OffsetDateTime endTimeMax, Boolean isActive,
+        Integer skip, Integer top);
 
     /**
      * Lists the step executions of a job execution.
@@ -107,8 +107,8 @@ public interface JobStepExecutionsClient {
     @ServiceMethod(returns = ReturnType.COLLECTION)
     PagedIterable<JobExecutionInner> listByJobExecution(String resourceGroupName, String serverName,
         String jobAgentName, String jobName, UUID jobExecutionId, OffsetDateTime createTimeMin,
-        OffsetDateTime createTimeMax, OffsetDateTime endTimeMin, OffsetDateTime endTimeMax, Boolean isActive, Long skip,
-        Long top, Context context);
+        OffsetDateTime createTimeMax, OffsetDateTime endTimeMin, OffsetDateTime endTimeMax, Boolean isActive,
+        Integer skip, Integer top, Context context);
 
     /**
      * Gets a step execution of a job execution.
